refactor toDot logic to Renderer class
